feat: 알림 응답 서비스 수정

+++ src/main/java/com/example/feedprep/domain/notification/service/NotificationServiceImpl.java
@@ -106,7 +106,7 @@ public class NotificationServiceImpl implements NotificationService{
 	}
 
 	@Override
-	public Integer getNotificationCount(Long receiverId) {
+	public Long getNotificationCount(Long receiverId) {
 		return notificationRepository.getCount();
 	}
 }
